SoC a1 : kernel/freertos: fix the coverity detection. [2/4]
PD#SWPL-171211

Problem:
Detection ERROR CID 339578 338974 339313.

Solution:
Fixed the error.

Verify:
A113L_AD403

+++ bl30/rtos_sdk/kernel/freertos/aml_extend/aml_dmalloc_ext.c
@@ -61,7 +61,7 @@ void vdRecordFree(size_t xFreeSize)
 	}
 }
 
-int vPrintDmallocInfo(size_t tid)
+int vPrintDmallocInfo(int tid)
 {
 	int ret = 1;
